🌟 [Major]: Introducing the `Hashtable` module (#4)
## Description

This pull request introduces the `Hashtable` module.

Some functions have moved here from `Utilities` and `GitHub`.

Added functions:

*
[`src/functions/public/ConvertFrom-Hashtable.ps1`](diffhunk://#diff-b0c173d6838b6ee74826974d3bc58c4e9865d24dec0a636fc6266cd3eb94363bR1-R88):
Added a new filter to convert a hashtable to a PSCustomObject, making it
easier to work with structured data.
*
[`src/functions/public/ConvertTo-Hashtable.ps1`](diffhunk://#diff-3ec07d2a54272155b3304518b07c92497a86e4d28544a4ad3ed62fdae5f3d1a4R1-R92):
Introduced a filter to convert an object to a hashtable, useful for
serialization purposes.
*
[`src/functions/public/Format-Hashtable.ps1`](diffhunk://#diff-7d90dfc5bc21b6d23ceed48105316975686cabff108964a18fa84407c1429f87R1-R128):
Added a filter to convert a hashtable to its PowerShell code
representation, aiding in exporting configurations.
*
[`src/functions/public/Merge-Hashtable.ps1`](diffhunk://#diff-85f850d9ef5b00c3b57cf8ab3108f346c8b931130ca45d535fdde27369dd7983R1-R110):
Implemented a filter to merge multiple hashtables, with support for
override sequences and a force option.
*
[`src/functions/public/Remove-HashtableEntry.ps1`](diffhunk://#diff-93a75f4c299fda32b547771ab30920f31120006910377202b64b478982aedfe3R1-R111):
Added a filter to remove specific entries from a hashtable based on
value, type, or name, with enhanced keep and remove conditions.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R120):
Updated the module name and description, added detailed usage examples,
and improved the readability and structure of the documentation.

Removal of outdated functions:

*
[`src/functions/public/New-PSModuleTest.ps1`](diffhunk://#diff-90b55db370b273157a3df0db4b895fca9740d25094d604894640fb0c5b220d9cL1-L35):
Removed the `New-PSModuleTest` function and its aliases, which were no
longer relevant to the module's purpose.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas