🩹 [Patch]: Handle empty hashtable in `Format-Hashtable` function (#5)
## Description

This pull request includes changes to the `Format-Hashtable` function
and its corresponding tests to handle empty hashtables correctly. The
most important changes are as follows:

Improvements to `Format-Hashtable` function:

*
[`src/functions/public/Format-Hashtable.ps1`](diffhunk://#diff-7d90dfc5bc21b6d23ceed48105316975686cabff108964a18fa84407c1429f87R68-R72):
Added a check to return '@{}' immediately if the hashtable is empty.

Enhancements to tests:

*
[`tests/Hashtable.Tests.ps1`](diffhunk://#diff-0d9a4623edf86d4025f7602749bc2c1397bec49ddcbb1fa7b51fe34634049687R258-R267):
Added a new test context to ensure that the `Format-Hashtable` function
returns an empty hashtable string when given an empty hashtable.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas